+++ include/Matrix2x2.h
@@ -4,16 +4,16 @@
 #include <complex>
 
 /**
- *  A 2x2 matrix of std::complex<float> numbers.
+ *  A 2x2 matrix of std::complex<double> numbers.
  *  (lazy implementation: only required functions are defined)
  */
 struct Matrix2x2 {
-    std::complex<float> e11;
-    std::complex<float> e12;
-    std::complex<float> e21;
-    std::complex<float> e22;
+    std::complex<double> e11;
+    std::complex<double> e12;
+    std::complex<double> e21;
+    std::complex<double> e22;
 
-    Matrix2x2(std::complex<float> e11_, std::complex<float> e12_, std::complex<float> e21_, std::complex<float> e22_)
+    Matrix2x2(std::complex<double> e11_, std::complex<double> e12_, std::complex<double> e21_, std::complex<double> e22_)
         : e11(e11_), e12(e12_), e21(e21_), e22(e22_) {}
     
     Matrix2x2(const Matrix2x2& other)
@@ -25,10 +25,10 @@ struct Matrix2x2 {
 };
 
 struct Vector2 {
-    std::complex<float> e1;
-    std::complex<float> e2;
+    std::complex<double> e1;
+    std::complex<double> e2;
 
-    Vector2(std::complex<float> e1_, std::complex<float> e2_)
+    Vector2(std::complex<double> e1_, std::complex<double> e2_)
         : e1(e1_), e2(e2_) {}
     
     Vector2(const Vector2& other)
