+++ tests/Propagation.cpp
@@ -4,8 +4,8 @@
 #include "../Field.h"
 
 int main(int argc, char* argv[]) {
-    float angleOfIncidence = M_PI / 6;
-    float E1inc = 1;
+    double angleOfIncidence = M_PI / 6;
+    double E1inc = 1;
 
     Field f = Field({ 10 }, { 1 });
     f.
